refactor: rename method and modify directory structure
